+++ src/test/java/org/logic2j/core/SolutionApiTest.java
@@ -261,7 +261,7 @@ public class SolutionApiTest extends PrologTestBase {
         final String goal = "perm([a,b,c,d,e,f,g,h], Q)";
         final GoalHolder holder = getProlog().solve(goal);
         ProfilingInfo.setTimer1();
-        final long count = holder.count();
+        final int count = holder.count();
         ProfilingInfo.reportAll("Number of solutions to " + goal + " is " + count);
         assertThat(count).isEqualTo(40320);
     }
